iniziata l'implementazione del controllo delle autorizzazioni sui metodi

+++ src/main/java/TAASS/ServiceDBEventi/controllers/EventoController.java
@@ -1,6 +1,7 @@
 package TAASS.ServiceDBEventi.controllers;
 
 import TAASS.ServiceDBEventi.classiComode.IscriviEvento;
+import TAASS.ServiceDBEventi.exception.MyCustomException;
 import TAASS.ServiceDBEventi.models.Evento;
 import TAASS.ServiceDBEventi.repositories.EventoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,6 +10,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -17,24 +19,42 @@ import java.util.*;
 @RestController
 @RequestMapping("/api/v1/evento")
 public class EventoController {
-    /*TODO:
-        -eventi a cui partecipa una persona
-        -eventi di un tipo
-        -eventi di un tipo in un comune
-     */
+    //x-auth-user-role  ROLE_ADMIN, ROLE_CLIENT, ROLE_MAYOR, ROLE_PUBLISHER
 
     @Autowired
     private EventoRepository eventoRepository;
 
     @GetMapping
-    public List<Evento> getAllEventi(){
+    public List<Evento> getAllEventi(HttpServletRequest requestHeader){
+        /*System.out.println("**************************************************************************************************************");
+
+        Enumeration headerNames = requestHeader.getHeaderNames();
+        while (headerNames.hasMoreElements()) {
+            String key = (String) headerNames.nextElement();
+            String value = requestHeader.getHeader(key);
+            System.out.println("HEADER: " + key +" " + value);
+        }*/
+
+
+        /*if(requestHeader.getHeader("x-auth-user-role").equals("ROLE_ADMIN")){
+            System.out.println("Sei autorizzato ad accedere a questo metodo");
+        }else{
+            throw new MyCustomException("FORBIDDEN", HttpStatus.FORBIDDEN);
+        }*/
+        //Auth: solo admin
+        if(!requestHeader.getHeader("x-auth-user-role").equals("ROLE_ADMIN")){
+            //Solo un admin può guardare tutti gli eventi di tutti, anche quelli scaduti
+            throw new MyCustomException("FORBIDDEN", HttpStatus.FORBIDDEN);
+        }
+
         List<Evento> eventi = new ArrayList<>();
         eventoRepository.findAll().forEach(eventi::add);
         return eventi;
     }
 
     @GetMapping("/non-scaduti")
-    public List<Evento> getEventiNonScaduti(){
+    public List<Evento> getEventiNonScaduti(HttpServletRequest requestHeader){
+        //Auth: all
         List<Evento> eventi;
         Date date =ottieniData();
         System.out.println("# eventi non scaduti: data attuale: " + date);
@@ -44,6 +64,7 @@ public class EventoController {
 
     @GetMapping("/eventi-non-iscritto/{utenteID}")
     public List<Evento> getEventiUtenteNonIscritto(@PathVariable long utenteID ){
+        //TODO: SERVE?
         System.out.println("# getEventiUtenteNonIscritto: uid: " + utenteID);
         List<Evento> eventi = eventoRepository.findEventiUtenteNonIscritto(utenteID);
         System.out.println("# getEventiUtenteNonIscritto: size: " + eventi.size());
@@ -51,11 +72,15 @@ public class EventoController {
     }
 
     @GetMapping("/eventi-non-iscritto-non-scaduti/{utenteID}")
-    public List<Evento> getEventiUtenteNonIscrittoNonscaduti(@PathVariable long utenteID ){
+    public List<Evento> getEventiUtenteNonIscrittoNonscaduti(HttpServletRequest requestHeader, @PathVariable long utenteID ){
+        //Auth: solo l'utente con id = utenteID
+        if(Integer.parseInt(requestHeader.getHeader("x-auth-user-id")) != utenteID){
+            throw new MyCustomException("FORBIDDEN", HttpStatus.FORBIDDEN);
+        }
         System.out.println("# getEventiUtenteNonIscritto: uid: " + utenteID);
         //start debug
         //stampa tutti gli iscritti a tutti gli eventi
-        List<Evento> tuttiEventi = getAllEventi();
+        List<Evento> tuttiEventi = getAllEventi(requestHeader);
         System.out.println("#\tlista iscrizioni eventi");
         for(int i = 0; i < tuttiEventi.size(); i++){
             System.out.println("#\t\tevento: " + tuttiEventi.get(i).getId() + ", nome = " + tuttiEventi.get(i).getNome());
@@ -78,21 +103,23 @@ public class EventoController {
 
     @GetMapping("/nome/{nome}")
     public List<Evento> trovaPerNome(@PathVariable String nome){
+        //Auth: tutti
         List <Evento> eventi = eventoRepository.findByNome(nome);
         return eventi;
     }
 
     @GetMapping("/tipo/{tipo}")
     public List<Evento> trovaPerTipo(@PathVariable long tipo){
+        //Auth: tutti
         System.out.println("# trovaPerTipo: tipo richiesto: " + tipo);
         List<Evento> eventi = eventoRepository.findByTipoEventoId(tipo);
         System.out.println("# trovaPerTipo: elementi trovati: " + eventi.size());
         return eventi;
     }
 
-    //si può cancellare
     @GetMapping("/info-evento/{id}")
     public Evento trovaPerID(@PathVariable long id){
+        //Auth: tutti
         System.out.println("# trovaPerID: id = " + id);
         Optional<Evento> evento = eventoRepository.findById(id);
         if(evento.isPresent()){
@@ -106,7 +133,29 @@ public class EventoController {
     }
 
     @GetMapping("/iscritti-evento/{id}")
-    public HashSet<Long> iscrittiEvento(@PathVariable long id){
+    public HashSet<Long> iscrittiEvento(@PathVariable long id,HttpServletRequest requestHeader){
+        //Auth: Admin, sindaco del comune dell'evento e pubblicatore dell'evento
+        switch (requestHeader.getHeader("x-auth-user-role")){
+            case "ROLE_ADMIN":{
+                //può accedere
+                break;
+            }
+            case "ROLE_CLIENT":{
+                //non può accedere a questo
+                throw new MyCustomException("FORBIDDEN", HttpStatus.FORBIDDEN);
+            }
+            case "ROLE_MAYOR":{
+                //può accedere solo se l'evento è del comune del sindaco
+                break;
+            }
+            case "ROLE_PUBLISHER":{
+                //può accedere solo se è il publisher dello stesso evento
+                break;
+            }
+            default:{
+                throw new MyCustomException("FORBIDDEN", HttpStatus.FORBIDDEN);
+            }
+        }
         Optional<Evento> iscritti = eventoRepository.findById(id);
         if(iscritti.isPresent()){
             return (HashSet<Long>) iscritti.get().getIscritti();
@@ -118,6 +167,9 @@ public class EventoController {
 
     @GetMapping("/eventi-comune/{comune}")
     public List<Evento> eventiDelComune(@PathVariable long comune){
+        //permette di vedere anche gli eventi scaduti
+        //Auth: sindaco e anche publisher
+
         System.out.println("# richiesta di tutti gli eventi del comune: " + comune);
         List<Evento> eventi = eventoRepository.findByComune(comune);
         return eventi;
@@ -125,27 +177,40 @@ public class EventoController {
 
     @GetMapping("/eventi-comune-non-scaduti/{comune}")
     public List<Evento> eventiDelComuneNonScaduti(@PathVariable long comune){
+        //Auth: tutti
         System.out.println("# richiesta di tutti gli eventi del comune: " + comune);
         List<Evento> eventi = eventoRepository.findByComuneNonScaduti(comune, ottieniData());
         return eventi;
     }
 
     @GetMapping("/iscrizione-utente/{utenteID}")
-    public List<Evento> eventiUtenteIscritto(@PathVariable long utenteID){
+    public List<Evento> eventiUtenteIscritto(@PathVariable long utenteID, HttpServletRequest requestHeader){
+        //Auth: utente con id = utenteID
         //restituisce l'elenco di tutti gli eventi a cui è iscritto un utente
+        if(Integer.parseInt(requestHeader.getHeader("x-auth-user-id")) != utenteID){
+            throw new MyCustomException("FORBIDDEN", HttpStatus.FORBIDDEN);
+        }
         return eventoRepository.findIscrizioniUtente(utenteID);
     }
 
     @GetMapping("/iscrizione-utente-non-scaduti/{utenteID}")
-    public List<Evento> eventiUtenteIscrittoNonScaduti(@PathVariable long utenteID){
+    public List<Evento> eventiUtenteIscrittoNonScaduti(@PathVariable long utenteID, HttpServletRequest requestHeader){
+        //Auth: utente con id = utenteID
         //restituisce l'elenco di tutti gli eventi a cui è iscritto un utente
+        if(Integer.parseInt(requestHeader.getHeader("x-auth-user-id")) != utenteID){
+            throw new MyCustomException("FORBIDDEN", HttpStatus.FORBIDDEN);
+        }
         return eventoRepository.findIscrizioniUtenteNonScaduti(utenteID, ottieniData());
     }
 
     @PostMapping("/disiscrivi")
-    public ResponseEntity<Map<String,String>> disiscrivi(@RequestBody Map<String,Long> body){
+    public ResponseEntity<Map<String,String>> disiscrivi(@RequestBody Map<String,Long> body, HttpServletRequest requestHeader){
+        //Auth: utente con id = utenteID
         Evento evento = trovaPerID(body.get("evento_id"));
         long utenteID = body.get("utente_id");
+        if(Integer.parseInt(requestHeader.getHeader("x-auth-user-id")) != utenteID){
+            throw new MyCustomException("FORBIDDEN", HttpStatus.FORBIDDEN);
+        }
         Map<String,String> response = new HashMap<>();
         if(evento.getIscritti().contains(utenteID)){
             evento.getIscritti().remove(utenteID);
@@ -160,13 +225,19 @@ public class EventoController {
     }
 
     @DeleteMapping("/deleteAll")
-    public ResponseEntity<String> rimuoviTuttiEventi(){
+    public ResponseEntity<String> rimuoviTuttiEventi(HttpServletRequest requestHeader){
+        //Auth: admin
+        if(!requestHeader.getHeader("x-auth-user-role").equals("ROLE_ADMIN")){
+            //Solo un admin può guardare tutti gli eventi di tutti, anche quelli scaduti
+            throw new MyCustomException("FORBIDDEN", HttpStatus.FORBIDDEN);
+        }
         eventoRepository.deleteAll();
         return new ResponseEntity<>("tutti gli eventi sono stati cancellati con successo", HttpStatus.OK);
     }
 
     @DeleteMapping("/deleteById/{eventoID}")
     public ResponseEntity<Map<String, String>> eliminaEventoID(@PathVariable long eventoID/*@PathVariable long  id*/){
+        //Admin: sindaco del comune dell'evento e pubblicatore dell'evento
         Map<String, String> response = new HashMap<String, String>();
         if(trovaPerID(eventoID) != null){
             eventoRepository.deleteById(eventoID);
@@ -180,6 +251,7 @@ public class EventoController {
 
     @PostMapping
     public Evento addEvento(@RequestBody Evento evento){
+        //Auth: admin, sindaco e publisher del comune
         //@DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
         System.out.println("# Aggiungi evento");
         System.out.println("#\taddEvento: aggiungo evento: " + evento);
@@ -194,6 +266,7 @@ public class EventoController {
 
     @PostMapping("/iscrivi")
     public ResponseEntity<Evento> iscrivi(@RequestBody IscriviEvento iscriviEvento){
+        //Auth: utente con id = utenteid
         System.out.println("# iscrivi: Ho ricevuto: evento = " + iscriviEvento.getEvento() + "; utente = " + iscriviEvento.getUtente());
         Optional<Evento> evento = eventoRepository.findById(iscriviEvento.getEvento());
         try {
@@ -209,6 +282,7 @@ public class EventoController {
     //TODO: verificare se si può creare in maniera più "bella"
     @PutMapping("/aggiorna/{id}")
     public ResponseEntity<Evento> aggiornaEvento(@PathVariable("id") long id, @RequestBody Evento evento){
+        //Auth: admin, sindaco del comune, publisher proprietario
         System.out.println("# aggiornaEvento: ricevuto id = " + id);
         Optional<Evento> datiEvento = eventoRepository.findById(id);
 
@@ -236,6 +310,7 @@ public class EventoController {
         }
     }
 
+    //TODO: cancellare
     @PostMapping("/prenota")
     public ResponseEntity<Map<String, String>> prenotaEvento(@RequestBody Map<String, Long> body){
         //verifico che ci siano ancora posti disponibili
